Fix internal seed # off by 1
The internal seed # displayed on the Tracker UI and used for the Next Seed feature was 1 lower than it should have been.

In effect, the seed # used should match exactly to the seed # present in the seed.txt file. This is more intuitive for the user to understand and to change if needed.